+++ sentry/src/main/java/io/sentry/JsonSerializer.java
@@ -105,7 +105,7 @@ public final class JsonSerializer implements ISerializer {
     deserializersByClass.put(SentryStackFrame.class, new SentryStackFrame.Deserializer());
     deserializersByClass.put(SentryStackTrace.class, new SentryStackTrace.Deserializer());
     deserializersByClass.put(
-        SentryStartupProfilingOptions.class, new SentryStartupProfilingOptions.Deserializer());
+        SentryAppStartProfilingOptions.class, new SentryAppStartProfilingOptions.Deserializer());
     deserializersByClass.put(SentryThread.class, new SentryThread.Deserializer());
     deserializersByClass.put(SentryTransaction.class, new SentryTransaction.Deserializer());
     deserializersByClass.put(Session.class, new Session.Deserializer());